updated zaino-state structure
